refactor: clean up code everywhere pt.1
