+++ src/main.rs
@@ -12,7 +12,7 @@
 //      The rest deal with each app screen individually.
 
 mod data;
-mod menu;
+mod menu_screen;
 mod player_select;
 mod score_table;
 mod winner_screen;
@@ -53,7 +53,7 @@ fn app(cx: Scope) -> Element {
     };
 
     match state.read().screen {
-        Screen::Menu => cx.render(rsx!(menu::intro_screen())),
+        Screen::Menu => cx.render(rsx!(menu_screen::main_menu())),
         Screen::PlayerSelect => cx.render(rsx!(player_select::player_select())),
         Screen::Game => cx.render(rsx!(score_table::score_table())),
         Screen::Winner => cx.render(rsx!(winner_screen::winner_screen())),
